Parse language prefix paths like we had them on the old site
This tries to be clever about reducing boiler plate.
It also fixes the way we identified the languates. Since on Transifex we
used the lang_COUNTRY format, we must follow this here as well, paying
attention that the coutry abbreviation must be upper case.

This also uses the session object to persist the chosen language. This
should make the website do the right thing, regarless whether the user
comes in with the generic URL and we parse the browser language
preference, or whether the user explicitly asks for a language.

I expect this to be more fine tuning, but it seems to be a good start.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>