add GitHub webhook support
This requires the webhook secret to verify the validity of the call.
It also stores the release_ids that it sees in the env so that we
don't forget to process them in case the server process gets restarted.

The code tries to make sure that only one worker actually deals with the
assets and updates the env (which then automagically updates the latest
release website). Both when called from the webhook, and when restarted
and finding unprocessed release_ids.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>