list both language and language_COUNTRY as supported languages
Otherwise if a user lists nl,en;q=0.9,nl_NL;q=0.8 as preferred languages
(we can argue if this makes a log of sense, but apparently that's what
Edge on Windows does), will get en instead of nl_NL - which is likely
not what they expect.

Thanks to Peter Zaal for pointing out the underlying issue and helping
me figure out this solution.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>